% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{give_headspace}
\alias{give_headspace}
\title{Give Headspace}
\usage{
give_headspace(selector = "nav", session = NULL)
}
\arguments{
\item{selector}{CSS select to apply headroom to.}

\item{session}{A valid shiny session.}
}
\description{
Initialise headspace server-side.
}
