% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{use_headspace}
\alias{use_headspace}
\title{Include cicerone dependencies in your Shiny UI.}
\usage{
use_headspace()
}
\description{
Include cicerone dependencies in your Shiny UI.
}
\examples{
library(shiny)

ui <- fluidPage(
 use_headspace()
)

server <- function(input, output){}

if(interactive()) shinyApp(ui, server)

}
